src/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
celery==5.3.4
redis==5.0.1
aiogram==3.3.0
requests==2.31.0
django-cors-headers==4.3.1
